got geocoding working

+++ app/models.py
@@ -50,8 +50,8 @@ class Porch(db.Document):
     location = db.ReferenceField(Location)
     time_available_start = db.DateTimeField(default=datetime.utcnow)
     time_available_end = db.DateTimeField(default=datetime.utcnow)
-    lat = db.FloatField()
-    long = db.FloatField()
+    lat = db.DecimalField()
+    long = db.DecimalField()
     meta = {
         'indexes': [
             {'fields': ('address', 'location'), 'unique': True}
